fix sanitize filename
  the length limit of filename is 255 bytes(not characters)
